+++ src/main/java/com/debatetimer/dto/timebased/request/TimeBasedTableCreateRequest.java
@@ -1,4 +1,4 @@
-package com.debatetimer.dto.time_based.request;
+package com.debatetimer.dto.timebased.request;
 
 import com.debatetimer.domain.TimeBoxes;
 import com.debatetimer.domain.member.Member;
