+++ src/main/java/com/debatetimer/dto/timebased/response/TimeBasedTimeBoxResponse.java
@@ -1,4 +1,4 @@
-package com.debatetimer.dto.time_based.response;
+package com.debatetimer.dto.timebased.response;
 
 import com.debatetimer.domain.Stance;
 import com.debatetimer.domain.timebased.TimeBasedBoxType;
