+++ src/main/java/com/debatetimer/repository/timebased/TimeBasedTimeBoxRepository.java
@@ -1,4 +1,4 @@
-package com.debatetimer.repository.time_based;
+package com.debatetimer.repository.timebased;
 
 import com.debatetimer.domain.TimeBoxes;
 import com.debatetimer.domain.timebased.TimeBasedTable;
