+++ src/test/java/com/debatetimer/controller/BaseDocumentTest.java
@@ -12,7 +12,7 @@ import com.debatetimer.exception.errorcode.ClientErrorCode;
 import com.debatetimer.service.auth.AuthService;
 import com.debatetimer.service.member.MemberService;
 import com.debatetimer.service.parliamentary.ParliamentaryService;
-import com.debatetimer.service.time_based.TimeBasedService;
+import com.debatetimer.service.timebased.TimeBasedService;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.filter.log.RequestLoggingFilter;
