+++ src/test/java/com/debatetimer/service/BaseServiceTest.java
@@ -9,8 +9,8 @@ import com.debatetimer.fixture.TimeBasedTimeBoxGenerator;
 import com.debatetimer.repository.member.MemberRepository;
 import com.debatetimer.repository.parliamentary.ParliamentaryTableRepository;
 import com.debatetimer.repository.parliamentary.ParliamentaryTimeBoxRepository;
-import com.debatetimer.repository.time_based.TimeBasedTableRepository;
-import com.debatetimer.repository.time_based.TimeBasedTimeBoxRepository;
+import com.debatetimer.repository.timebased.TimeBasedTableRepository;
+import com.debatetimer.repository.timebased.TimeBasedTimeBoxRepository;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
