+++ src/test/java/com/debatetimer/service/timebased/TimeBasedServiceTest.java
@@ -1,4 +1,4 @@
-package com.debatetimer.service.time_based;
+package com.debatetimer.service.timebased;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,10 +9,10 @@ import com.debatetimer.domain.member.Member;
 import com.debatetimer.domain.timebased.TimeBasedBoxType;
 import com.debatetimer.domain.timebased.TimeBasedTable;
 import com.debatetimer.domain.timebased.TimeBasedTimeBox;
-import com.debatetimer.dto.time_based.request.TimeBasedTableCreateRequest;
-import com.debatetimer.dto.time_based.request.TimeBasedTableInfoCreateRequest;
-import com.debatetimer.dto.time_based.request.TimeBasedTimeBoxCreateRequest;
-import com.debatetimer.dto.time_based.response.TimeBasedTableResponse;
+import com.debatetimer.dto.timebased.request.TimeBasedTableCreateRequest;
+import com.debatetimer.dto.timebased.request.TimeBasedTableInfoCreateRequest;
+import com.debatetimer.dto.timebased.request.TimeBasedTimeBoxCreateRequest;
+import com.debatetimer.dto.timebased.response.TimeBasedTableResponse;
 import com.debatetimer.exception.custom.DTClientErrorException;
 import com.debatetimer.exception.errorcode.ClientErrorCode;
 import com.debatetimer.service.BaseServiceTest;
